refactoring row level components
