+++ packages/react-bootstrap-table2/src/row/row-section.js
